Split DefaultControllerTest file
